+++ keil/motor.h
@@ -12,7 +12,7 @@ sbit Px2 = P3^6;
 sbit Px3 = P3^7; //步进电机控制引脚
 
 void motor_data(uchar da);
-void motor_angle(uchar angle);
+void motor_angle(int angle);
 void motor_direction(uchar D,int S);
 
 #endif
